+++ ShopAdmin/src/main/java/com/vision/shoppingmall/product/model/request/CreateProductRequest.java
@@ -9,10 +9,10 @@ public class CreateProductRequest {
   private String publisherName;
   private String authorName;
   private String translatorName;
-  private Double purchasePrice;
-  private Double unitPrice;
-  private Double discountPrice;
-  private Double sellingPrice;
+  private int purchasePrice;
+  private int unitPrice;
+  private int discountPrice;
+  private int sellingPrice;
   private String thumbnailImage;
   private String productImage;
 }
